travel_telegram_bot/requirements.txt
python-telegram-bot==22.7
python-dotenv==1.1.1
pytest==9.0.2
